Use randomized ID for file ID to avoid conflict (#3)
Currently, when a youtube video is downloaded, the file name is based on
the video ID. This may be problematic in the future as 2 requests sent
in are asking for the same youtube video but with different configs
(resolution, output extension, etc). This commit aims to fix this by
generating a random ID as the file name instead to avoid conflicts.